refactor: move execution and process to internal
