+++ packages/core/src/Ignition.ts
@@ -2,11 +2,11 @@ import type {
   DeploymentResult,
   IgnitionDeployOptions,
   UpdateUiAction,
-} from "./types/deployment";
+} from "./internal/types/deployment";
 import type {
   ExecutionResultsAccumulator,
   ExecutionVisitResult,
-} from "./types/executionGraph";
+} from "./internal/types/executionGraph";
 import type { ICommandJournal } from "./types/journal";
 import type { Module, ModuleDict } from "./types/module";
 import type { IgnitionPlan } from "./types/plan";
@@ -17,13 +17,13 @@ import type {
 
 import setupDebug from "debug";
 
-import { Deployment } from "./deployment/Deployment";
-import { execute } from "./execution/execute";
-import { loadJournalInto } from "./execution/loadJournalInto";
-import { hashExecutionGraph } from "./execution/utils";
+import { Deployment } from "./internal/deployment/Deployment";
+import { execute } from "./internal/execution/execute";
+import { loadJournalInto } from "./internal/execution/loadJournalInto";
+import { hashExecutionGraph } from "./internal/execution/utils";
+import { generateDeploymentGraphFrom } from "./internal/process/generateDeploymentGraphFrom";
+import { transformDeploymentGraphToExecutionGraph } from "./internal/process/transformDeploymentGraphToExecutionGraph";
 import { NoopCommandJournal } from "./journal/NoopCommandJournal";
-import { generateDeploymentGraphFrom } from "./process/generateDeploymentGraphFrom";
-import { transformDeploymentGraphToExecutionGraph } from "./process/transformDeploymentGraphToExecutionGraph";
 import { Services } from "./types/services";
 import { IgnitionError } from "./utils/errors";
 import { resolveProxyValue } from "./utils/proxy";
