+++ packages/core/src/internal/execution/dispatch/executeDeployedContract.ts
@@ -1,11 +1,11 @@
+import type { Services } from "../../../types/services";
 import type {
   DeployedContract,
   ExecutionResultsAccumulator,
   ExecutionVertexVisitResult,
 } from "../../types/executionGraph";
-import type { Services } from "../../types/services";
 
-import { VertexResultEnum } from "../../types/graph";
+import { VertexResultEnum } from "../../../types/graph";
 
 import { resolveFrom, toAddress } from "./utils";
