+++ packages/core/src/internal/execution/execute.ts
@@ -1,4 +1,5 @@
-import type { Deployment } from "../deployment/Deployment";
+import type { Deployment } from "../../internal/deployment/Deployment";
+import type { Services } from "../../types/services";
 import type { ExecutionOptions, ExecutionState } from "../types/deployment";
 import type { ExecutionVertexDispatcher } from "../types/execution";
 import type {
@@ -7,10 +8,9 @@ import type {
   ExecutionVisitResult,
   ExecutionVertex,
 } from "../types/executionGraph";
-import type { Services } from "../types/services";
 
-import { viewExecutionResults } from "../deployment/utils";
-import { IgnitionError } from "../utils/errors";
+import { viewExecutionResults } from "../../internal/deployment/utils";
+import { IgnitionError } from "../../utils/errors";
 
 import { ExecutionGraph } from "./ExecutionGraph";
 import { executionDispatch } from "./dispatch/executionDispatch";
