+++ packages/core/src/internal/process/generateDeploymentGraphFrom.ts
@@ -2,12 +2,12 @@ import type {
   CallPoints,
   DeploymentBuilderOptions,
   IDeploymentGraph,
-} from "../types/deploymentGraph";
+} from "../../types/deploymentGraph";
 
-import { DeploymentBuilder } from "../dsl/DeploymentBuilder";
-import { Module, ModuleDict } from "../types/module";
-import { IgnitionError } from "../utils/errors";
-import { assertModuleReturnTypes } from "../utils/guards";
+import { DeploymentBuilder } from "../../dsl/DeploymentBuilder";
+import { Module, ModuleDict } from "../../types/module";
+import { IgnitionError } from "../../utils/errors";
+import { assertModuleReturnTypes } from "../../utils/guards";
 
 export function generateDeploymentGraphFrom<T extends ModuleDict>(
   ignitionModule: Module<T>,
