+++ packages/core/src/internal/process/transformDeploymentGraphToExecutionGraph.ts
@@ -1,12 +1,12 @@
-import { ExecutionGraph } from "../execution/ExecutionGraph";
-import { clone } from "../graph/adjacencyList";
+import { clone } from "../../graph/adjacencyList";
 import {
   IDeploymentGraph,
   DeploymentGraphVertex,
-} from "../types/deploymentGraph";
+} from "../../types/deploymentGraph";
+import { Services } from "../../types/services";
+import { ExecutionGraph } from "../execution/ExecutionGraph";
 import { ExecutionVertex, IExecutionGraph } from "../types/executionGraph";
 import { TransformResult } from "../types/process";
-import { Services } from "../types/services";
 
 import { convertDeploymentVertexToExecutionVertex as convertDeploymentVertexToExecutionVertex } from "./transform/convertDeploymentVertexToExecutionVertex";
 import { reduceDeploymentGraphByEliminatingVirtualVertexes } from "./transform/reduceDeploymentGraphByEliminatingVirtualVertexes";
