+++ packages/core/src/internal/types/process.ts
