+++ packages/core/test/deploymentBuilder/accounts.ts
@@ -8,7 +8,7 @@ import { assert } from "chai";
 import { ethers } from "ethers";
 
 import { buildModule } from "../../src/dsl/buildModule";
-import { generateDeploymentGraphFrom } from "../../src/process/generateDeploymentGraphFrom";
+import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
 import { isCall, isHardhatContract } from "../../src/utils/guards";
 
 import { getDeploymentVertexByLabel } from "./helpers";
