+++ packages/core/test/deploymentBuilder/buildModule.ts
@@ -2,7 +2,7 @@
 import { assert } from "chai";
 
 import { buildModule } from "../../src/dsl/buildModule";
-import { generateDeploymentGraphFrom } from "../../src/process/generateDeploymentGraphFrom";
+import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
 import { IgnitionError } from "../../src/utils/errors";
 
 describe("deployment builder - buildModule", () => {
