+++ packages/core/test/deploymentBuilder/calls.ts
@@ -2,7 +2,7 @@
 import { assert } from "chai";
 
 import { buildModule } from "../../src/dsl/buildModule";
-import { generateDeploymentGraphFrom } from "../../src/process/generateDeploymentGraphFrom";
+import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
 import {
   IDeploymentBuilder,
   IDeploymentGraph,
