+++ packages/core/test/deploymentBuilder/sendETH.ts
@@ -3,7 +3,7 @@ import { assert } from "chai";
 import { ethers } from "ethers";
 
 import { buildModule } from "../../src/dsl/buildModule";
-import { generateDeploymentGraphFrom } from "../../src/process/generateDeploymentGraphFrom";
+import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
 import {
   IDeploymentBuilder,
   IDeploymentGraph,
