+++ packages/core/test/deploymentBuilder/value.ts
@@ -9,7 +9,7 @@ import { assert } from "chai";
 import { ethers } from "ethers";
 
 import { buildModule } from "../../src/dsl/buildModule";
-import { generateDeploymentGraphFrom } from "../../src/process/generateDeploymentGraphFrom";
+import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
 import {
   isArtifactContract,
   isHardhatContract,
